requirements.txt
flask
mistralai
python-dotenv
elevenlabs
